moved the schema file
