+++ server/src/database/schemas/index.ts
@@ -6,4 +6,4 @@ export * from './department.schema';
 export * from './faculty.schema';
 export * from './classroom.schema';
 export * from './specialty.schema';
-export * from './file.schema';
+export * from '../../files/file.schema';
