+++ server/src/files/files.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { MongooseModule } from '@nestjs/mongoose';
 import { FilesController } from './files.controller';
 import { FilesService } from './files.service';
-import { File, FileSchema } from '../database/schemas/file.schema';
+import { File, FileSchema } from './file.schema';
 import { AuditLogModule } from '../audit-log/audit-log.module';
 
 @Module({
